pyproject.toml
[project]
name = "ai-exam-evaluator"
version = "0.1.0"
description = "AI Exam Evaluator: Automated grading system with rubric alignment."
authors = [
    { name = "Vitthal Chandane", email = "vitthal@example.com" }
]
license = "MIT"

# Dependencies required for your environment
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv",
    "openenv-core",
    "pydantic"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
